201411182249
Originally checked in by Hujolopter\Sebastian on 2014-11-18 22:50:12.457 as Changeset 67

git-tfs-id: [https://infovision.visualstudio.com/DefaultCollection]$/RoughSets;C103

+++ RoughSets/Infovision.Datamining.RoughSet/Bireduct.cs
@@ -16,18 +16,18 @@ namespace Infovision.Datamining.Roughset
 
         #region Constructors
         
-        public Bireduct(DataStore dataStore, int[] fieldIds, int[] objectIndexes, double approxDegree)
+        public Bireduct(DataStore dataStore, int[] fieldIds, int[] objectIndexes, int approxDegree)
             : base(dataStore, fieldIds, approxDegree)
         {
             objectSet = new ObjectSet(dataStore, objectIndexes);            
         }
 
-        public Bireduct(DataStore dataStore, int[] fieldIds, double approxDegree)
+        public Bireduct(DataStore dataStore, int[] fieldIds, int approxDegree)
             : this(dataStore, fieldIds, new int[] { }, approxDegree)
         {
         }
 
-        public Bireduct(DataStore dataStore, double approxDegree)
+        public Bireduct(DataStore dataStore, int approxDegree)
             : this(dataStore, dataStore.DataStoreInfo.GetFieldIds(FieldTypes.Standard), new int[] { }, approxDegree)
         {
         }
