+++ RoughSets/Infovision.Datamining.RoughSet/ReductWeight.cs
@@ -12,18 +12,18 @@ namespace Infovision.Datamining.Roughset
     {
         #region Constructors
 
-        public ReductWeights(DataStore dataStore, int[] fieldIds, double[] objectWeights, double approximationDegree)
+        public ReductWeights(DataStore dataStore, int[] fieldIds, double[] objectWeights, int approximationDegree)
             : base(dataStore, fieldIds, approximationDegree)
         {
             this.Weights = objectWeights;
         }
 
-        public ReductWeights(DataStore dataStore, int[] fieldIds, double approximationDegree)
+        public ReductWeights(DataStore dataStore, int[] fieldIds, int approximationDegree)
             : this(dataStore, fieldIds, null, approximationDegree)
         {
         }
 
-        public ReductWeights(DataStore dataStore, double approximationDegree)
+        public ReductWeights(DataStore dataStore, int approximationDegree)
             : this(dataStore, new int[] { }, approximationDegree)
         {
         }
